[SymForce] Move things to typing_util
As suggested on dataclass tuple ops PR

Topic: sym-typing-util
Relative: sym-dataclass-ops-tuples
Reviewers: hayk,bradley
GitOrigin-RevId: 5587fd947e05f2618c6c4ef47cad805148672035
