+++ symforce/ops/impl/dataclass_group_ops.py
@@ -6,12 +6,8 @@
 import dataclasses
 
 from symforce.ops import GroupOps
-from symforce.python_util import (
-    get_type,
-    get_sequence_from_dataclass_sequence_field,
-    maybe_tuples_of_types_from_annotation,
-)
 from symforce import typing as T
+from symforce import typing_util
 
 from .dataclass_storage_ops import DataclassStorageOps
 
@@ -25,12 +21,12 @@ class DataclassGroupOps(DataclassStorageOps):
             for field in dataclasses.fields(a):
                 field_type = type_hints_map[field.name]
                 if field.metadata.get("length") is not None:
-                    sequence_instance = get_sequence_from_dataclass_sequence_field(
+                    sequence_instance = typing_util.get_sequence_from_dataclass_sequence_field(
                         field, field_type
                     )
                     constructed_fields[field.name] = GroupOps.identity(sequence_instance)
                 elif (
-                    sequence_types := maybe_tuples_of_types_from_annotation(field_type)
+                    sequence_types := typing_util.maybe_tuples_of_types_from_annotation(field_type)
                 ) is not None:
                     # It's a Tuple of known size
                     constructed_fields[field.name] = GroupOps.identity(sequence_types)
@@ -40,21 +36,21 @@ class DataclassGroupOps(DataclassStorageOps):
         else:
             for field in dataclasses.fields(a):
                 constructed_fields[field.name] = GroupOps.identity(getattr(a, field.name))
-            return get_type(a)(**constructed_fields)
+            return typing_util.get_type(a)(**constructed_fields)
 
     @staticmethod
     def compose(a: T.Dataclass, b: T.Dataclass) -> T.Dataclass:
-        assert get_type(a) == get_type(b)
+        assert typing_util.get_type(a) == typing_util.get_type(b)
         constructed_fields = {}
         for field in dataclasses.fields(a):
             constructed_fields[field.name] = GroupOps.compose(
                 getattr(a, field.name), getattr(b, field.name)
             )
-        return get_type(a)(**constructed_fields)
+        return typing_util.get_type(a)(**constructed_fields)
 
     @staticmethod
     def inverse(a: T.Dataclass) -> T.Dataclass:
         constructed_fields = {}
         for field in dataclasses.fields(a):
             constructed_fields[field.name] = GroupOps.inverse(getattr(a, field.name))
-        return get_type(a)(**constructed_fields)
+        return typing_util.get_type(a)(**constructed_fields)
