+++ symforce/ops/impl/dataclass_storage_ops.py
@@ -6,12 +6,8 @@
 import dataclasses
 
 from symforce.ops import StorageOps
-from symforce.python_util import (
-    get_type,
-    get_sequence_from_dataclass_sequence_field,
-    maybe_tuples_of_types_from_annotation,
-)
 from symforce import typing as T
+from symforce import typing_util
 
 
 class DataclassStorageOps:
@@ -42,12 +38,12 @@ class DataclassStorageOps:
             for field in dataclasses.fields(a):
                 field_type = type_hints_map[field.name]
                 if field.metadata.get("length") is not None:
-                    sequence_instance = get_sequence_from_dataclass_sequence_field(
+                    sequence_instance = typing_util.get_sequence_from_dataclass_sequence_field(
                         field, field_type
                     )
                     count += StorageOps.storage_dim(sequence_instance)
                 elif (
-                    sequence_types := maybe_tuples_of_types_from_annotation(field_type)
+                    sequence_types := typing_util.maybe_tuples_of_types_from_annotation(field_type)
                 ) is not None:
                     # It's a Tuple of known size
                     count += StorageOps.storage_dim(sequence_types)
@@ -76,7 +72,7 @@ class DataclassStorageOps:
             for field in dataclasses.fields(a):
                 field_type = type_hints_map[field.name]
                 if field.metadata.get("length") is not None:
-                    sequence_instance = get_sequence_from_dataclass_sequence_field(
+                    sequence_instance = typing_util.get_sequence_from_dataclass_sequence_field(
                         field, field_type
                     )
                     storage_dim = StorageOps.storage_dim(sequence_instance)
@@ -84,7 +80,7 @@ class DataclassStorageOps:
                         sequence_instance, elements[offset : offset + storage_dim]
                     )
                 elif (
-                    sequence_types := maybe_tuples_of_types_from_annotation(field_type)
+                    sequence_types := typing_util.maybe_tuples_of_types_from_annotation(field_type)
                 ) is not None:
                     # It's a Tuple of known size
                     storage_dim = StorageOps.storage_dim(sequence_types)
@@ -108,7 +104,7 @@ class DataclassStorageOps:
                     field_instance, elements[offset : offset + storage_dim]
                 )
                 offset += storage_dim
-            return get_type(a)(**constructed_fields)
+            return typing_util.get_type(a)(**constructed_fields)
 
     @staticmethod
     def symbolic(a: T.DataclassOrType, name: T.Optional[str], **kwargs: T.Dict) -> T.Dataclass:
@@ -128,14 +124,16 @@ class DataclassStorageOps:
                 field_type = type_hints_map[field.name]
                 try:
                     if field.metadata.get("length") is not None:
-                        sequence_instance = get_sequence_from_dataclass_sequence_field(
+                        sequence_instance = typing_util.get_sequence_from_dataclass_sequence_field(
                             field, field_type
                         )
                         constructed_fields[field.name] = StorageOps.symbolic(
                             sequence_instance, f"{name_prefix}{field.name}", **kwargs
                         )
                     elif (
-                        sequence_types := maybe_tuples_of_types_from_annotation(field_type)
+                        sequence_types := typing_util.maybe_tuples_of_types_from_annotation(
+                            field_type
+                        )
                     ) is not None:
                         # It's a Tuple of known size
                         constructed_fields[field.name] = StorageOps.symbolic(
@@ -149,7 +147,7 @@ class DataclassStorageOps:
                     raise NotImplementedError(
                         f"Could not create field {field.name} of type {field_type}"
                     ) from ex
-            return get_type(a)(**constructed_fields)
+            return typing_util.get_type(a)(**constructed_fields)
         else:
             constructed_fields = {}
             name_prefix = f"{name}." if name is not None else ""
@@ -163,4 +161,4 @@ class DataclassStorageOps:
                     raise NotImplementedError(
                         f"Could not create field {field.name} of type {field_instance}"
                     ) from ex
-            return get_type(a)(**constructed_fields)
+            return typing_util.get_type(a)(**constructed_fields)
