+++ symforce/ops/impl/sequence_lie_group_ops.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from symforce.ops import StorageOps
 from symforce.ops import LieGroupOps
-from symforce.python_util import get_type
+from symforce.typing_util import get_type
 from symforce import typing as T
 
 from .sequence_group_ops import SequenceGroupOps
