+++ symforce/ops/lie_group_ops.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import symforce.internal.symbolic as sf
 from symforce import typing as T
 from symforce import logger
-from symforce.python_util import get_type
+from symforce.typing_util import get_type
 
 from .group_ops import GroupOps
