+++ symforce/ops/storage_ops.py
@@ -3,7 +3,7 @@
 # This source code is under the Apache 2.0 license found in the LICENSE file.
 # ----------------------------------------------------------------------------
 
-from symforce.python_util import get_type, scalar_like
+from symforce.typing_util import get_type, scalar_like
 import symforce.internal.symbolic as sf
 from symforce import typing as T
